Simplify the release process
This patch includes a couple of things to make it easier to do a
new GitHub release for this repository.

First, the pattern match for git release tags was made more stringent.

Second, a new pipeline was written, which is triggered by a release
tag. This pipeline automatically creates a new GitHub release, based
on the tag name.

Finally, a simple shell script was created to trigger releases. It
works by pushing a new release tag, based on the current code version
stored in pom.xml.