Made the scheduler more generic
I removed the dependency of the job scheduler on Directory sync, so
it could theoretically be used in other projects.

It can either run a job once or repeatedly, depending on the timer
settings.

To make this work, I had to stop passing parameters around in method
signatures. I switched to a solution that pulls the information it
needs directly from the settings file.