Generalized support for diagnosis code correction
Previously, only the star model was able to correct ICD 10 codes to make
them suitable for the Directory. This push generalizes this so that
it also works for the regular Directory update.

Additionally, the star model update has been made into an optional
feature.

The README has been updated to reflect the new functionality.