Incorporated the directory-sync library into the code
Took the code from https://github.com/samply/directory-sync and added
it to this repository. This means that there is no dependency with
the directory-sync repository any more. The original rational behind
the separation was that the library might be used elsewhere, not just
in a standalone service. This never actually happened, hence the
decision to combine service and library, which makes the code easier
to maintain.