Added new functionality to allow Directory sync to work with GraphQL API
The primary changes have been in DirectoryApi, DirectoryCalls and
DirectoryEndpoints, which have become abstract classes, with different
implementations depending on the APIs they are connected to (either the
original REST API or the new GraphQL API).

Sync has been modified to first test if a GraphQL API is available and
use it if so. Otherwise, it tries to find a REST API. I.e. no additional
configuration is required on the part of the user.

Minor changes were needed elsewhere to impedance match between the data
representation needed by REST and the data representation needed by
GraphQL.