New GraphQL database auto select mechanism
The old onew relied on trying GraphQL endpoints until one actually
worked. The problem with this approach is that it caused Molgenis to
break in such a way that it froze as soon as it was given a correct
endpoint.

So instead, I switched to pulling a list of databases from Molgenis
and then using a selection heuristic for picking the right one.

Along the way, a number of methods became redundant and I removed them.