Improved fact table generation
Petr Holub wanted to make sure that the output from the Java fact
table generator in this Directory sync implementation is the same as
the output from his R script.

To make comparisons easier, a file output for Directory sync was
implemented, where tables in semi-colon separated format are generated.
These were be compared with the output from the R program and it
was discovered that the patient age at first diagnosis was different.

The problem was fixed and now both the R-script and Directory sync
produce exactly the same fact tables, even for large numbers of patients
(tested with circa 11000 patients).