Relegated the DirectoryCollectionPut object to the DirectoryApi
This object is really meant as a template for the PUT operation for
collections in the RESTful API. It doesn't really belong in the
higher parts of the architecture. They shouldn't know about that
kind of thing.

Instead, I have taken the Collection class and expanded it to be able
to deal with all of the attributes in a collection, then used it
as the universal currency for communicating collection information.